// mobile/gen/models/CreateTriage.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const createTriage201StatusEnum = {
  PENDENTE: "PENDENTE",
  EM_ANALISE: "EM_ANALISE",
  CANCELADO: "CANCELADO",
  REENCAMINHADO: "REENCAMINHADO",
} as const;

export type CreateTriage201StatusEnumKey =
  (typeof createTriage201StatusEnum)[keyof typeof createTriage201StatusEnum];

/**
 * @description triage created successfully
 */
export type CreateTriage201 = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  complaint: string;
  /**
   * @type string
   */
  symptomDuration: string;
  /**
   * @type string
   */
  symptom: string;
  /**
   * @type string
   */
  actionTaken: string;
  /**
   * @type string
   */
  reactionAfterAction: string;
  /**
   * @default "PENDENTE"
   * @type string
   */
  status: CreateTriage201StatusEnumKey;
  /**
   * @type string
   */
  patientId: string;
  /**
   * @type string | undefined
   */
  analyzedById?: string;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  updatedAt: string;
};

/**
 * @description Invalid input data provided
 */
export type CreateTriage400 = any;

/**
 * @description Authentication required
 */
export type CreateTriage401 = any;

/**
 * @description Insufficient permissions
 */
export type CreateTriage403 = any;

export const createTriageMutationRequestStatusEnum = {
  PENDENTE: "PENDENTE",
  EM_ANALISE: "EM_ANALISE",
  CANCELADO: "CANCELADO",
  REENCAMINHADO: "REENCAMINHADO",
} as const;

export type CreateTriageMutationRequestStatusEnumKey =
  (typeof createTriageMutationRequestStatusEnum)[keyof typeof createTriageMutationRequestStatusEnum];

export type CreateTriageMutationRequest = {
  /**
   * @minLength 1
   * @type string
   */
  complaint: string;
  /**
   * @minLength 1
   * @type string
   */
  symptomDuration: string;
  /**
   * @minLength 1
   * @type string
   */
  symptom: string;
  /**
   * @minLength 1
   * @type string
   */
  actionTaken: string;
  /**
   * @minLength 1
   * @type string
   */
  reactionAfterAction: string;
  /**
   * @type string | undefined
   */
  status?: CreateTriageMutationRequestStatusEnumKey;
  /**
   * @type object | undefined
   */
  patient?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @type object | undefined
   */
  analyzedBy?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
};

export type CreateTriageMutationResponse = CreateTriage201;

export type CreateTriageMutation = {
  Response: CreateTriage201;
  Request: CreateTriageMutationRequest;
  Errors: CreateTriage400 | CreateTriage401 | CreateTriage403;
};