// mobile/gen/models/FindTriages.ts
export const dataStatusEnum2 = {
  PENDENTE: "PENDENTE",
  EM_ANALISE: "EM_ANALISE",
  CANCELADO: "CANCELADO",
  REENCAMINHADO: "REENCAMINHADO",
} as const;

export type DataStatusEnum2Key =
  (typeof dataStatusEnum2)[keyof typeof dataStatusEnum2];

/**
 * @description List of triages retrieved successfully
 */
export type FindTriages200 = {
  /**
   * @description Total number of records matching the criteria
   * @type integer | undefined
   */
  total?: number;
  /**
   * @description Number of records returned in current page
   * @type integer | undefined
   */
  results?: number;
  /**
   * @type array | undefined
   */
  data?: {
    /**
     * @type string
     */
    id: string;
    /**
     * @type string
     */
    complaint: string;
    /**
     * @type string
     */
    symptomDuration: string;
    /**
     * @type string
     */
    symptom: string;
    /**
     * @type string
     */
    actionTaken: string;
    /**
     * @type string
     */
    reactionAfterAction: string;
    /**
     * @default "PENDENTE"
     * @type string
     */
    status: DataStatusEnum2Key;
    /**
     * @type string
     */
    patientId: string;
    /**
     * @type string | undefined
     */
    analyzedById?: string;
    /**
     * @type string, date-time
     */
    createdAt: string;
    /**
     * @type string, date-time
     */
    updatedAt: string;
  }[];
};